REFACTOR: Architecting app.py with blueprint
