+++ static/index.css
@@ -1,0 +1,103 @@
+@import url('https://fonts.googleapis.com/css2?family=Inconsolata:wght@200;600;900&family=Noto+Sans+TC:wght@100;300;400;500;700;900&display=swap');
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+html,
+body {
+    font-family: 'Noto Sans TC', sans-serif;
+    font-size: 16px;
+    /* position: relative; */
+}
+
+/* header */
+.navBar {
+    width: 100%;
+    height: 54px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    background: #650606;
+    display: flex;
+    justify-content: center;
+
+}
+
+.navBarContainer {
+    width: 1200px;
+    display: flex;
+    top: 0;
+    justify-content: space-between;
+    align-items: center;
+    outline: 1px solid red;
+
+}
+
+.navBarTpeText {
+    font-weight: 700;
+    font-size: 30px;
+    line-height: 34px;
+    color: #448899;
+}
+
+.navBarBtn {
+    padding: 20px;
+    display: flex;
+    justify-content: end;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 13px;
+    color: #666666;
+}
+
+.navBarBtnItinerary {
+    margin-right: 20px;
+}
+
+.navBarBtnLogin {
+    margin-left: 20px;
+}
+
+
+/* footer */
+footer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 40px 843px;
+
+    position: absolute;
+    width: 1920px;
+    height: 104px;
+    left: 0px;
+    top: 1270px;
+
+    /* Gray 700 */
+
+    background: #757575;
+}
+
+.footerText {
+    position: absolute;
+    width: 234px;
+    height: 14px;
+    /* left: 0px;
+    top: 5px; */
+
+    /* Body - Bold 16(13.3) */
+
+    font-style: normal;
+    font-weight: 700;
+    line-height: 13px;
+    /* identical to box height, or 83% */
+
+    display: flex;
+    align-items: center;
+
+    /* Additional Color/White */
+
+    color: #FFFFFF;
+}
