+++ templates/index test.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Taipei Day Trip</title>
+    <link rel="stylesheet" href="/static/index.css">
+</head>
+
+<body>
+    <!-- header -->
+    <nav class="navBar">
+        <div class="navBarContainer">
+            <div class="navBarTpe">
+                <p class="navBarTpeText">台北一日遊</p>
+            </div>
+            <div class="navBarBtn">
+                <div class=".navBarBtnItinerary">預定行程</div>
+                <div class="navBarBtnLogin">登入/註冊</div>
+            </div>
+        </div>
+
+    </nav>
+    <header>
+        <!-- navBar -->
+
+        <!-- nav -->
+        <nav></nav>
+    </header>
+
+
+    <!-- main -->
+    <main>
+        <!-- section -->
+        <section></section>
+    </main>
+
+
+    <!-- footer -->
+    <footer>
+        <p class="footerText">COPYRIGHT © 2021 台北一日遊</p>
+    </footer>
+</body>
+
+</html>
